Fix issues with how targets are stored in MultiTask.
Previously we had simple solution to use target_dict[task_name] to store
the target dict for that task.

But then if we want to run augmentations, they expect the data to appear in
top-level of the target-dict rather than like target_dict[task_name][key].

So now it is stored like target_dict[task_name + key] with a prefix. Then the
whole thing can be specified in the augmentation.

Maybe it'd be better to support the augmentations directly in MultiTask? In
multisat we seamlessly handled the augmentations but that's because we knew
exactly how all the different labels were going to be used...