Support layers specifying an alias for storing in the tile store.
This is useful if you have multiple layers with same data source but
need to define the layers separately instead of using max_matches > 0.
Then you can set them all to have same alias so the items from the
data source don't need to be ingested multiple times.

Probably there's a way we can just save them based on the data source
name but for now we need to manually set this alias.