Use KML file for Sentinel-2 tile lookup instead of mgrs library.
This removes a dependency but the main goal is to fix problem where sometimes
an MGRS cell is not being returned properly during the lookup process. This is
because previously we iterate over longitude/latitude in 0.5 increment and
lookup the cell ID at each position. This works for normal-sized cells (which
are about 100 km x 100 km) but there are some smaller cells at the boundary
of UTM zones that can be skipped.

So now we instead create grid index over KML containing Sentinel-2 tiles and
look it up that way. The KML is converted to JSON, and JSON data is cached in
a cache directory that data source must provide.