Store items in the tile store under the source projection instead of destination projection.
Resolves #63.

Originally I think the items were re-projected to the destination projection (i.e., the
projection(s) of windows that require the item) because we were building a big satellite
image dataset and we were reading the images directly from the tile store and wanted
them in the target projection.

Re-projecting during ingestion saves time to re-project during materialization, but if
the windows have a much higher resolution than the source data, but only cover a small
portion of the item, then it wastes a lot of space. Space is also wasted if there are
many different windows with different projections referencing the same item, since
currently we store one copy of the item per destination projection in the tile store.

So this change makes it store the item as-is (except rasters are converted to tiled
GeoTIFF and some data sources do things like Sentinel-2 harmonization), in the source
projection, and then uses rasterio.vrt to re-project when materializing the window.

This also improves correctness because rasterio.vrt is more likely to handle the
re-projection correctly than our code. Like we could have bug if the CRS is already
correct but the GeoTIFF comes with a sub-pixel offset, we would just round it off I
think but rasterio.vrt will do the resampling as needed.