add class_names/class_property_name to dataset config file.
these are for annotation / prediction layers to indicate which GeoJSON properties
correspond to some notion of a class, or for non-image GeoTIFFs to define what the
different values mean. it could be useful in the future to do things like rasterize
a vector layer to a corresponding raster layer. for now it will be used for es
studio syncing.