Improve the API for RasterFormat and VectorFormat.
RasterFormat: the previous API accepts a projection and bounds when encoding, but
only bounds when decoding. This commit changes it to also accept projection when
decoding, and the RasterFormat is expected to re-project the data to the requested
projection. Previously, the caller had to do extra work to ensure the bounds
provided exactly matches the projection of the actual raster. Now, if the caller
wants data in a different projection they can pass it rather than re-projecting
themselves.

VectorFormat: the previous API accepts a projection when encoding (no bounds),
and bounds when decoding (no projection). Encoding is changed to accept neither
projection nor bounds: neither are necessary since the Features include a
projection. Decoding is changed to accept both a projection and bounds for similar
reasons to RasterFormat.

The original motivation of this change was because of issues employing
TileVectorFormat for the vector format used in DefaultTileStore, since the API
didn't really support a way for windows in different projections to ask for
portions of the vector data.

I think the new API makes a lot more sense because (a) passing around bounds
without a corresponding projection is strange especially when the formats don't
necessarily have a way to get the projection of the stored data, and (b) there
was already some re-projecting logic in DefaultTileStore which now is just moved
to VectorFormat/RasterFormat.