Fast geometry to_projection when dst crs matches src crs.
The main motivation was to increase the speed of rtree index creation in the
gcp_public_data data source. It seems rasterio.warp.transform_geom takes 10+
ms even when the src/dst CRS are the same, so skipping the call saves a lot
of time when we are doing many re-projections.