Speedup LocalFiles data source by caching the list of items.
Previously the items are computed every time the LocalFiles is instantiated which
could take many minutes if there are lots of source files since it neesd to check
the bounds of each source file.

Now this still needs to be done, but its result will be cached so future instantiations
will be fast.