Fix issues using custom samplers with multi-GPU training.
Previously, when setting a sampler in SplitConfig with multi-GPU training, the sampler
would just get overwritten by pytorch Lightning since it would configure a
DistributedSampler. When using rslearn_projects, it happens even with single-GPU
training since we currently always enable ddp to set find_unused_parameters.

This leads to confusing bugs where you expect it to use e.g. random weighted sampling
but the option is effectively ignored since Lightning configures a different sampler.

Now the DistributedSampler replacement functionality is disabled, and we instead
configure the sampler appropriately in the data module even when sampler factory is
not set.