Fix bug where prediction fails due to trying to use task_weights.
I also moved the task_weights from MultiTaskModel to RslearnLightningModule so
it can be applied for different kinds of models and also for individual losses
within a task instead of just across tasks (e.g. object detection returns
multiple losses that can be weighted separately).

I added integration test for this that fails without the changes. I also fixed
a bug with ClassificationTask.process_outputs when read_class_id is set, where
it tries to write a numpy scalar as JSON which fails instead of converting it
to Python data type first.

Resolves #225.