GEE: don't treat no valid pixels export error as fatal.
This happens during materialization when a window geometry intersects a
GEE Image, but the Image does not actually have any valid pixels at that
location. If there are multiple items in the item group, this is problematic
because it means the materialization fails for that window rather than
getting no data from that item but potentially getting valid pixels from
other items.

So it is changed to treat this error as okay and just return no data.