If load_all_layers is set, require all the layers to be present.
Otherwise, we will have batches of examples where each example has time
series of a different length, and the models won't work.

In the future, we may want this to be a separate option, so you can stack
data from whatever layers are available. Currently, though, all the models
require consistent batches so it makes more sense to just connect load_all_layers
with needing all the configured layers to be present.