Fix read error when read is not aligned for lower zoom offsets.
An error occurs if we try to read a patch that is not a multiple of the
offset factor, e.g. if it is stored at zoom offset -2 (1/4 of the window
resolution) but the requested bounds is not a multiple of 4.

This fixes it by identifying when the read is not aligned properly and in
that case use rasterio.vrt instead of doing the zooming in/out. We still
use nearest neighbor interpolation.

The only reason to not always do this method is because zooming in/out
may be faster, but this needs to be benchmarked. If it is the same speed,
it would be simpler to just rely on rasterio.vrt.