Fix bug with AllPatchesDataset/RslearnWriter where empty output may be written.
If the number of data loader workers exceeds the number of windows, AllPatchesDataset
will distribute the same window to multiple workers. This is okay for validation and
testing, but for prediction, the RslearnWriter will not be able to distinguish the
partial patch outputs from different data loader workers, and so if the workers yield
the window in parallel, and both provide the last patch at the same time, then the
last patch from the first worker will be processed correctly, but the last patch from
the second worker will find no other patch outputs and thus overwrite the output layer
with empty data.

We now require the number of workers to be <= the number of windows, and the
functionality to distribute the same window to multiple workers is removed.