Fix occasional re-projection error when reading vector features.
The error seems to sometimes happen when a feature in a GeoJSON falls outside the
area of use of the requested projection. It doesn't always happen and I'm not sure
what causes the exact error conditions.

Anyway, the solution employed here is to check in WGS84 and use that to determine
if we should even attempt the re-projection to the destination projection. In WGS84
we do need to deal with antimeridian splitting but otherwise should work okay.