Allows template parameters in dataset config.json (#309)
In the same vein as #278 + #283, this changeset allows
template variables of the format `${TEMPLATE_VAR}` into
dataset config.jsons, which will be overwritten with
injectable env vars at startup. It leverages the same logic
that was used in model.yaml's argparser, but applied
during Dataset construction.

The changes facilitate esrun's need to to coordinate the writing
of label layers before fine tuning and targeting of prediction
output layers during inference into a discoverable location.