Resolves pystac memleak (hopefully) (#335)
Our Planetary Computer data sources use the [pystac](https://pystac.readthedocs.io/en/stable/index.html)
library as an HTTP client to retrieve and normalize data
from Microsoft's API.

This library is designed to build up object graphs / caches in-memory
over the lifetime of a process and does not appear to be intended
for use in long-running applications. There is no way to opt out of,
cap the growth of, or clear the cache. Every pystac object, including
the client, is registered as a node in the object graph, and that
graph gets added to whenever results resolve from HTTP responses.

The consequence is that when running dataset handler operations
through our planetary computer DSes, we are experiencing unbounded
memory growth. In my local tests, this could be as much as 100MB/min
per worker process. Previously we worked around this by running handler
operations in a multiproc pool with limited executions per worker process.

This changeset instead sets a max-Items budget for a pystac `Client`
object. When this budget is exceeded, the Client is replaced with a fresh
client. Due to the stickiness of the multi-way references on pystac graph nodes,
I found it was necessary to do two additional things to make the client and its friends
eligible for garbage collection:

* clear members off the Client node (`root.clear_links()` etc)
* stop using `Collections` and `CollectionClients`

On the latter -- the collection client we were using previously was mostly
sugar around an HTTP request we can perform directly off the parent
client in `get_by_name`. Note this will use the search/ endpoint instead
of the collection/ endpoint now, but according to
https://planetarycomputer.microsoft.com/api/stac/v1
planetary computer is conformant to both.

Note, this is a breaking change specifically for the satlas data sources in RSLP,
which access but do not use the collection client previously initialized by
the Sentinel1/2 datasources. This will require a tiny update in rslp code, accordingly.

Memory growth before change:
<img width="790" height="172" alt="Screenshot 2025-10-15 at 12 58 09 PM" src="https://github.com/user-attachments/assets/55a4e7e7-c5e1-4cea-84b7-a0b2b748740e" />

Memory growth after change:
<img width="818" height="162" alt="Screenshot 2025-10-15 at 12 57 31 PM" src="https://github.com/user-attachments/assets/d831fa76-d93a-4172-b8ae-8af010ea3498" />

Note initial growth before GC has run enough times to start
ejecting long survivors, after which we reach equilibrium.