Fix bug with AllPatchesDataset when the window is smaller than the patch size.
Previously, when computing the patches, it would align with the upper end
of the window bounds, but it pads the window data expecting patches to go
beyond the upper end but not the lower end. So if the window is (0, 0, 2, 2)
with patch size 4, it would try to read patch (-2, -2, 2, 2) but the window
data does not get padded to support that.

Now instead the patch computing function is updated to ensure it returns
(0, 0, 4, 4) for that case, which works because we make sure the window
data is padded by patch_size on the right/bottom (upper end).